+++ src/main/kotlin/io/github/gunpowder/entities/StoredBalance.kt
@@ -1,0 +1,9 @@
+package io.github.gunpowder.entities
+
+import java.math.BigDecimal
+import java.util.*
+
+data class StoredBalance(
+    val uuid: UUID,
+    var balance: BigDecimal
+)